GSOC viewport project by Jason Wilkins. Master rebase work in progress.
Branch here still does not compile and there may be some changes to the
naming and API, but it should serve to keep things visible.

+++ source/blender/editors/include/BIF_glutil.h
@@ -206,10 +206,10 @@ void setlinestyle(int nr);
 /* own working polygon offset */
 void bglPolygonOffset(float viewdist, float dist);
 
-/* For caching opengl matrices (gluProject/gluUnProject) */
+/* For caching opengl matrices (gpuProject/gpuUnProject) */
 typedef struct bglMats {
-	double modelview[16];
-	double projection[16];
+	float modelview[16];
+	float projection[16];
 	int viewport[4];
 } bglMats;
 void bgl_get_mats(bglMats *mats);
